[team3_pair1] Q&A 3 구현했습니다. (#29)
* step3 implement

* feedback modify

+++ src/main/java/codesquad/service/QuestionService.java
@@ -2,6 +2,8 @@ package codesquad.service;
 
 import codesquad.domain.Question;
 import codesquad.domain.QuestionRepository;
+import codesquad.domain.User;
+import codesquad.exception.QuestionModifyFailException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -29,13 +31,19 @@ public class QuestionService {
         return questionRepository.findById(id).get();
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, User user) throws QuestionModifyFailException {
+        Question question = findById(id);
+        if(!question.isOwner(user))
+            throw new QuestionModifyFailException("자신의 글만 삭제할 수 있습니다.");
+
         questionRepository.deleteById(id);
     }
 
-    public void updateById(Long id, Question updateQuestion) {
+    public void updateById(Long id, Question updateQuestion, User user) throws QuestionModifyFailException {
         Question question = findById(id);
-        question.updateQuestion(updateQuestion);
+        if(!question.updateQuestion(updateQuestion,user))
+            throw new QuestionModifyFailException("자신의 글만 수정할 수 있습니다.");
+
         questionRepository.save(question);
     }
 }
